Combine queue test base classes. Remove duplication
