+++ src/test/java/net/openhft/chronicle/queue/QueueTestCommon.java
@@ -4,26 +4,88 @@ import net.openhft.chronicle.core.Jvm;
 import net.openhft.chronicle.core.OS;
 import net.openhft.chronicle.core.io.AbstractCloseable;
 import net.openhft.chronicle.core.io.AbstractReferenceCounted;
+import net.openhft.chronicle.core.io.IOTools;
 import net.openhft.chronicle.core.onoes.ExceptionKey;
 import net.openhft.chronicle.core.onoes.Slf4jExceptionHandler;
 import net.openhft.chronicle.core.threads.CleaningThread;
 import net.openhft.chronicle.core.threads.ThreadDump;
 import net.openhft.chronicle.core.time.SystemTimeProvider;
 import net.openhft.chronicle.wire.MessageHistory;
+import org.jetbrains.annotations.NotNull;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Rule;
+import org.junit.rules.ErrorCollector;
+import org.junit.rules.TestName;
+import org.junit.rules.TestRule;
+import org.junit.rules.TestWatcher;
+import org.junit.runner.Description;
 
+import java.io.File;
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 public class QueueTestCommon {
+    private static final boolean TRACE_TEST_EXECUTION = Jvm.getBoolean("queue.traceTestExecution");
+    private List<File> tmpDirs = new ArrayList<>();
     private final Map<Predicate<ExceptionKey>, String> ignoredExceptions = new LinkedHashMap<>();
     private final Map<Predicate<ExceptionKey>, String> expectedExceptions = new LinkedHashMap<>();
     protected ThreadDump threadDump;
     protected Map<ExceptionKey, Integer> exceptions;
     protected boolean finishedNormally;
+
+    static {
+        System.setProperty("queue.check.index", "true");
+    }
+
+    // *************************************************************************
+    // JUNIT Rules
+    // *************************************************************************
+
+    @Rule
+    public final TestName testName = new TestName();
+
+    @Rule
+    public final ErrorCollector errorCollector = new ErrorCollector();
+
+    @NotNull
+    @Rule
+    public TestRule watcher = new TestWatcher() {
+        @Override
+        protected void starting(@NotNull Description description) {
+            if (TRACE_TEST_EXECUTION) {
+                Jvm.debug().on(getClass(), "Starting test: "
+                        + description.getClassName() + "."
+                        + description.getMethodName()
+                );
+            }
+        }
+    };
+
+    // *************************************************************************
+    //
+    // *************************************************************************
+    static AtomicLong counter = new AtomicLong();
+
+    @BeforeClass
+    public static void synchronousFileTruncating() {
+        System.setProperty("chronicle.queue.synchronousFileShrinking", "true");
+    }
+
+    @NotNull
+    protected File getTmpDir() {
+        final String methodName = testName.getMethodName();
+        String name = methodName == null ? "unknown" : methodName;
+        final File tmpDir = DirectoryUtils.tempDir(name + "-" + counter.incrementAndGet());
+        tmpDirs.add(tmpDir);
+        return tmpDir;
+    }
 
     @Before
     public void assumeFinishedNormally() {
@@ -125,5 +187,11 @@ public class QueueTestCommon {
     }
 
     protected void tearDown() {
+        // should be able to remove tmp dirs
+        tmpDirs.forEach(file -> {
+            if (file.exists() && !IOTools.deleteDirWithFiles(file)) {
+                Jvm.error().on(getClass(), "Could not delete tmp dir " + file);
+            }
+        });
     }
 }
